monitor2: don't output extra spans
before this change, when a trace had a span that had a func that
matched a query, that trace suddenly would start getting all
spans recorded until the matching span was complete.
unfortunately, this could mean that we would pick up other spans
that had nothing to do with the presumably root span (perhaps
other span trees kicked off from earlier in the trace).

this change prunes the span tree to only return spans kicked off
by the matched span

Change-Id: I1e174627920626b38014559754f312b7bc56d2dd
Reviewed-on: https://gerrit.spacemonkey.com/12844
Tested-by: Jenkins CI <jenkins@spacemonkey.com>
Reviewed-by: Jeff Wendling <jeff@spacemonkey.com>