http: usage custom baggage headers as span annotations
This patch allows using "baggage: k1=v1,k2=v2" header, where the keys and values can be imported as span (assuming the key name is on the white list).

This feature can be used to tag/annotate distributed traces with the help of HTTP header, which makes it easier to find the right traces in the db (without using exemplars).